pyproject.toml
[project]
name = "cv-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.4",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langgraph-cli[inmem]>=0.3.6",
    "langgraph-prebuilt>=0.6.4",
    "langgraph-sdk>=0.2.0",
    "langsmith>=0.4.14",
    "notebook>=7.4.5",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
    "trustcall>=0.0.39",
    "wikipedia>=1.4.0",
]
